Remove explicit install of liblwgeom
This changeset removes the explicit installation of liblwgeom now that
it appears that the PostGIS package handles pulling in the version of
liblwgeom it depends on.

This avoids the runtime error from before that led to the explicit
installation of liblwgeom in this Ansible role. It also avoids a nasty
edge case where liblwgeom pulls in PostgreSQL 9.4 as a dependency
despite already having PostgreSQL 9.3 installed.